requirements.txt
torch 
numpy==1.24.2 
scikit-learn==1.0.1 
tqdm 
transformers==4.33.3 
datasets 
sentencepiece 
protobuf 
pytorch_lightning==1.2.1  
omegaconf 
hydra-core 
higher